tests: refactor: move `random_` helpers from tests.c to testutil.h
Can be reviewed via `--color-moved=dimmed-zebra`.
